Move mapif functions from int_clan.c to mapif.c
